migrate test_ddns to pytest
